# remove error log pre-allocation and add benchmark