--- a/retry_test.go
+++ b/retry_test.go
@@ -503,3 +503,29 @@ func TestUnwrap(t *testing.T) {
 	assert.Error(t, err)
 	assert.Equal(t, testError, errors.Unwrap(err))
 }
+
+func BenchmarkDo(b *testing.B) {
+	testError := errors.New("test error")
+
+	for i := 0; i < b.N; i++ {
+		Do(
+			func() error {
+				return testError
+			},
+			Attempts(10),
+			Delay(0),
+		)
+	}
+}
+
+func BenchmarkDoNoErrors(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		Do(
+			func() error {
+				return nil
+			},
+			Attempts(10),
+			Delay(0),
+		)
+	}
+}