This PR removes `errorLog` slice pre-allocation with the assumption that most calls to `retry.Do` don't actually require retry. If this is wishful thinking, please let me know. Aside from the readability win, the benchmarks that I added show improvement around allocations & cpu cycles when no errors are encountered, as expected.


Here are the results of the benchmarks that this PR introduces:
```sh
$ go test -benchmem -run=^$ -bench ^BenchmarkDo github.com/avast/retry-go/v4 > old.txt

goos: darwin
goarch: amd64
pkg: github.com/avast/retry-go/v4
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDo-16            	       3	 420393029 ns/op	    2197 B/op	      33 allocs/op
BenchmarkDoNoErrors-16    	 5136325	       230.7 ns/op	     368 B/op	       5 allocs/op
PASS
ok  	github.com/avast/retry-go/v4	4.331s
```

```sh
$ go test -benchmem -run=^$ -bench ^BenchmarkDo github.com/avast/retry-go/v4 > new.txt

goos: darwin
goarch: amd64
pkg: github.com/avast/retry-go/v4
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkDo-16            	       3	 372840109 ns/op	    2568 B/op	      38 allocs/op
BenchmarkDoNoErrors-16    	 7106366	       163.9 ns/op	     208 B/op	       4 allocs/op
PASS
ok  	github.com/avast/retry-go/v4	2.902s
```

```sh
$ benchstat old.txt new.txt

goos: darwin
goarch: amd64
pkg: github.com/avast/retry-go/v4
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
              │   old.txt    │                new.txt                │
              │    sec/op    │    sec/op     vs base                 │
Do-16           420.4m ± ∞ ¹   372.8m ± ∞ ¹        ~ (p=1.000 n=1) ²
DoNoErrors-16   230.7n ± ∞ ¹   163.9n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean         311.4µ         247.2µ        -20.62%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

              │    old.txt    │                new.txt                 │
              │     B/op      │     B/op       vs base                 │
Do-16           2.146Ki ± ∞ ¹   2.508Ki ± ∞ ¹        ~ (p=1.000 n=1) ²
DoNoErrors-16     368.0 ± ∞ ¹     208.0 ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean           899.2           730.9        -18.72%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

              │   old.txt   │               new.txt               │
              │  allocs/op  │  allocs/op   vs base                │
Do-16           33.00 ± ∞ ¹   38.00 ± ∞ ¹       ~ (p=1.000 n=1) ²
DoNoErrors-16   5.000 ± ∞ ¹   4.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean         12.85         12.33        -4.02%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```